server/mcp_server_vke/requirements.txt
mcp==1.12.0
pydantic>=2.11.3
python-dotenv>=1.1.0
retry>=0.9.2
starlette>=0.46.2
volcengine>=1.0.180
pyyaml>=6.0.2
kubernetes>=34.1.0